Core Gameplay Alpha 1 Changes to KayRahTayMan.java
KayRahTayMan.java has been modified to include the very core concepts of gameplay. In this version, a sprite can move left and right, and jump or double jump around the screen.

In order to work, there must be a new sprite named "redSquare.png" placed in your assets folder.

For me, this sprite is a placeholder, simply a red box 32x32 pixels.